Added release script for Linux, updated other Linux scripts to new standard, added chmod +x to launcher to avoid permission issues, and other minor fixes

+++ lin-build-launcher.sh
@@ -4,16 +4,13 @@ echo "BGArmor Executable Builder - Linux"
 echo "----------------------------------"
 echo
 
-# Get Python executable path from current platform configuration
-PYTHON_EXECUTABLE=$(cat ./launcher/Linux/python_executable.txt)
+source './source/tools/Linux/get-python.sh'
 
-# Resolve path variable
-eval $PYTHON_EXECUTABLE
+if [ ! -z $PYTHON_EXECUTABLE ]; then
+    $PYTHON_EXECUTABLE './source/tools/Common/helper_scripts/build_launcher.py'
+    rm -f Launcher.o
+fi
 
-# Execute script using given Python executable
-$PYTHON_EXECUTABLE './source/tools/Common/helper_scripts/build_launcher.py'
-
-rm -f Launcher.o
-
+echo
 echo "Done!"
 read -p "Press any key to continue..."
